requirements.txt
streamlit
langchain
langgraph
langsmith
qdrant-client
openai
pypdf
requests
python-dotenv
pytest
